+++ client/CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-<!-- Add unreleased changes here -->
+
+* Re-export `turnkey_api_key_stamper::TurnkeyP256ApiKey` from `turnkey_client`
 
 ## turnkey_client-v0.0.2